chore: remove needless file
