--- pubsub.go
@@ -1,15 +1,0 @@
-package rtfs
-
-import (
-	"errors"
-)
-
-// PubSubPublish is used to publish a a message to the given topic
-func (im *IpfsManager) PubSubPublish(topic string, data string) error {
-	if topic == "" {
-		return errors.New("topic is empty")
-	} else if data == "" {
-		return errors.New("data is empty")
-	}
-	return im.shell.PubSubPublish(topic, data)
-}
